+++ src/microservices/booking/graphql/input/index.ts
@@ -1,0 +1,1 @@
+export * from './create-booking.input';
