+++ src/microservices/booking/graphql/models/index.ts
@@ -1,0 +1,1 @@
+export * from './booking-gql.model';
