+++ src/microservices/booking/graphql/mutations/booking-mutation.resolver.ts
@@ -2,8 +2,8 @@ import { Args, Mutation, Resolver } from '@nestjs/graphql';
 import { UseGuards } from '@nestjs/common';
 
 import { GqlJwtAuthGuard } from '@auth/graphql/guards';
-import { GqlBooking } from '../models/booking-gql.model';
-import { CreateBookingInput } from '../input/create-booking.input';
+import { GqlBooking } from '../models';
+import { CreateBookingInput } from '../input/';
 import { BookingService } from '../../services';
 
 @Resolver((of) => GqlBooking)
