+++ src/microservices/booking/graphql/queries/booking-query.resolver.ts
@@ -3,7 +3,7 @@ import { Args, Resolver, Int, Query } from '@nestjs/graphql';
 
 import { GqlJwtAuthGuard } from '@auth/graphql/guards';
 import { BookingService } from '../../services';
-import { GqlBooking } from '../models/booking-gql.model';
+import { GqlBooking } from '../models';
 
 @Resolver((of) => GqlBooking)
 export class BookingQueryResolver {
